Changes to the create() method
create() initially created a record in the local collection, and then saved it to the server which sent down a slightly different record.
The server supplied record included an _id, which i've been using as the idAttribute.  This resulted in 2 nearly identical models in
the local collection and no way to reference the manually entered one.

Now, I don't add anything to the collection until it's provided by the server.  This makes create() async, which may result in problems,
but I don't have any ghosts living in my collection.